step_2/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
gunicorn==20.1.0
aiohttp==3.8.5
openai==0.28

pydantic~=2.5.2
python-dotenv~=1.0.0